Import `ActiveSupport::OptionMerger` implementation
The [ActiveSupport::OptionMerger][] is marked as `:nodoc:`, and is
intended to be private. It also coerces any `options` arguments into
keyword arguments by double-splatting them (with `**`).

This commit inlines that class' implementation into the
`AttributeMerger` class, and adds a special case for methods that accept
`options` as a final `Hash` argument, of which Action View (and Rails in
general) has many. When encountering a final `Hash`, pass the
`Attributes` instance instead, without splatting.

When the final argument is a "keyrest" like `**options`, splat the
`Attributes` instance into a `Hash`.

[ActiveSupport::OptionMerger]: https://github.com/rails/rails/blob/v7.0.4.1/activesupport/lib/active_support/option_merger.rb#L6